Termék lista elkészítésének folytatása 14.

+++ app/src/main/java/com/example/audioshopinventorymanagement/storagesscreen/StoragesScreenViewModel.kt
@@ -37,7 +37,7 @@ class StoragesScreenViewModel @Inject constructor(
         viewModelScope.launch(Dispatchers.IO) {
 
             if(arg != null){
-                val response = productApiRepository.getStoragesByWarehouseId(arg.toInt())
+                val response = productApiRepository.getStoragesByWarehouseId(arg)
 
                 when (response){
                     is ProductApiResponse.StoragesSuccess -> {
